frontend: Extract RouteSwitcher from App
