+++ frontend/src/App.tsx
@@ -9,38 +9,21 @@ import Toolbar from '@material-ui/core/Toolbar';
 import { SnackbarProvider } from 'notistack';
 import React from 'react';
 import { Provider } from 'react-redux';
-import { BrowserRouter, HashRouter, Redirect, Route, RouteProps, Switch } from 'react-router-dom';
+import { BrowserRouter, HashRouter } from 'react-router-dom';
+import RouteSwitcher from './components/App/RouteSwitcher';
 import TopBar from './components/App/TopBar';
 import { ClusterTitle } from './components/cluster/Chooser';
 import ActionsNotifier from './components/common/ActionsNotifier';
 import AlertNotification from './components/common/AlertNotification';
 import ReleaseNotes from './components/common/ReleaseNotes/ReleaseNotes';
-import Sidebar, { drawerWidth, NavigationTabs, useSidebarItem } from './components/Sidebar';
+import Sidebar, { drawerWidth, NavigationTabs } from './components/Sidebar';
 import helpers from './helpers';
 import { useElectronI18n } from './i18n/electronI18n';
 import ThemeProviderNexti18n from './i18n/ThemeProviderNexti18n';
-import { getToken } from './lib/auth';
-import { useClustersConf } from './lib/k8s';
-import { createRouteURL, getRoutePath, ROUTES } from './lib/router';
 import themes, { getThemeName, usePrefersColorScheme } from './lib/themes';
-import { getCluster } from './lib/util';
 import { initializePlugins } from './plugin';
 import { useTypedSelector } from './redux/reducers/reducers';
 import store from './redux/stores/store';
-
-function PageTitle({
-  title,
-  children,
-}: {
-  title: string | null | undefined;
-  children: React.ReactNode;
-}) {
-  React.useEffect(() => {
-    document.title = title || '';
-  }, [title]);
-
-  return <>{children}</>;
-}
 
 const useStyle = makeStyles(theme => ({
   root: {
@@ -70,44 +53,6 @@ const useStyle = makeStyles(theme => ({
     width: '1px',
   },
 }));
-
-function RouteSwitcher() {
-  const routes = useTypedSelector(state => state.ui.routes);
-
-  return (
-    <Switch>
-      {Object.values(ROUTES)
-        .concat(Object.values(routes))
-        .map((route, index) =>
-          route.name === 'OidcAuth' ? (
-            <Route
-              path={route.path}
-              component={() => (
-                <PageTitle title={route.name ? route.name : route.sidebar}>
-                  <route.component />
-                </PageTitle>
-              )}
-              key={index}
-            />
-          ) : (
-            <AuthRoute
-              key={index}
-              path={getRoutePath(route)}
-              sidebar={route.sidebar}
-              requiresAuth={!route.noAuthRequired}
-              requiresCluster={!route.noCluster}
-              exact={!!route.exact}
-              children={
-                <PageTitle title={route.name ? route.name : route.sidebar}>
-                  <route.component />
-                </PageTitle>
-              }
-            />
-          )
-        )}
-    </Switch>
-  );
-}
 
 function AppContainer() {
   const isSidebarOpen = useTypedSelector(state => state.ui.sidebar.isSidebarOpen);
@@ -192,52 +137,6 @@ function App() {
       </AppWithRedux>
     </Provider>
   );
-}
-
-interface AuthRouteProps {
-  children: React.ReactNode | JSX.Element;
-  sidebar: string | null;
-  requiresAuth: boolean;
-  requiresCluster: boolean;
-  [otherProps: string]: any;
-}
-
-function AuthRoute(props: AuthRouteProps) {
-  const { children, sidebar, requiresAuth = true, requiresCluster = true, ...other } = props;
-  const clusters = useClustersConf();
-  const redirectRoute = getCluster() ? 'login' : 'chooser';
-
-  useSidebarItem(sidebar);
-
-  function getRenderer({ location }: RouteProps) {
-    if (!requiresAuth) {
-      return children;
-    }
-
-    if (requiresCluster) {
-      const clusterName = getCluster();
-      if (!!clusterName) {
-        const cluster = clusters ? clusters[clusterName] : undefined;
-        const requiresToken = cluster?.useToken === undefined || cluster?.useToken;
-        if (!!getToken(clusterName) || !requiresToken) {
-          return children;
-        }
-      }
-    }
-
-    return (
-      <Redirect
-        to={{
-          pathname: createRouteURL(redirectRoute),
-          state: { from: location },
-        }}
-      />
-    );
-  }
-
-  // If no auth is required for the view, or the token is set up, then
-  // render the assigned component. Otherwise redirect to the login route.
-  return <Route {...other} render={getRenderer} />;
 }
 
 export default App;
